requirements.txt
instaloader
python-dotenv
requests
beautifulsoup4 